moved chared zod schemas, updated password generation for tests
