--- packages/wallet/backend/src/asset/validation.ts
@@ -1,8 +1,0 @@
-import { z } from 'zod'
-
-export const assetSchema = z.object({
-  body: z.object({
-    code: z.string(),
-    scale: z.number()
-  })
-})
