+++ packages/wallet/backend/src/auth/validation.ts
@@ -1,34 +1,14 @@
 import { z } from 'zod'
+import { emailSchema, loginSchema, signUpSchema } from '@wallet/shared'
 
-export const signUpSchema = z.object({
-  body: z
-    .object({
-      email: z.string().email({ message: 'Email is required' }),
-      password: z
-        .string()
-        .min(6, { message: 'Password should be at least 6 characters long' }),
-      confirmPassword: z.string()
-    })
-    .superRefine(({ password, confirmPassword }, ctx) => {
-      if (password !== confirmPassword) {
-        ctx.addIssue({
-          code: 'custom',
-          message: `Passwords do not match`,
-          path: ['confirmPassword']
-        })
-      }
-    })
+export const signUpBodySchema = z.object({
+  body: signUpSchema
 })
 
-export const logInSchema = z.object({
-  body: z.object({
-    email: z.string().email(),
-    password: z.string()
-  })
+export const logInBodySchema = z.object({
+  body: loginSchema
 })
 
-export const resendVerifyEmailSchema = z.object({
-  body: z.object({
-    email: z.string().email()
-  })
+export const emailBodySchema = z.object({
+  body: emailSchema
 })
