+++ packages/wallet/backend/src/rafiki/controller.ts
@@ -3,8 +3,8 @@ import { Logger } from 'winston'
 import { RatesService } from '@/rates/service'
 import { validate } from '@/shared/validate'
 import { RafikiService } from './service'
-import { ratesSchema, webhookSchema } from './validation'
-import { RatesResponse } from '@wallet/shared'
+import { webhookSchema } from './validation'
+import { ratesSchema, RatesResponse } from '@wallet/shared'
 
 interface IRafikiController {
   getRates: (
