--- packages/wallet/backend/src/rates/validation.ts
@@ -1,10 +1,0 @@
-import { z } from 'zod'
-
-export const ratesSchema = z.object({
-  query: z.object({
-    base: z
-      .string()
-      .length(3)
-      .transform((v) => v.toLocaleUpperCase())
-  })
-})
