+++ packages/wallet/backend/tests/mocks.ts
@@ -1,5 +1,5 @@
 import { faker } from '@faker-js/faker'
-import { logInSchema, signUpSchema } from '@/auth/validation'
+import { logInBodySchema, signUpBodySchema } from '@/auth/validation'
 import z from 'zod'
 import { PartialModelObject } from 'objection'
 import { Transaction } from '../src/transaction/model'
@@ -11,15 +11,16 @@ import {
   RapydProfile,
   walletSchema
 } from '@/rapyd/schemas'
-import { ratesSchema, webhookSchema } from '@/rafiki/validation'
+import { webhookSchema } from '@/rafiki/validation'
 import { EventType, WebHook } from '@/rafiki/service'
 import {
   incomingPaymentSchema,
   paymentDetailsSchema
 } from '@/incomingPayment/validation'
 import { outgoingPaymentSchema } from '@/outgoingPayment/validation'
+import { ratesSchema } from '@wallet/shared'
 
-export type LogInRequest = z.infer<typeof logInSchema>
+export type LogInRequest = z.infer<typeof logInBodySchema>
 export type GetRatesRequest = z.infer<typeof ratesSchema>
 export type OnWebHook = z.infer<typeof webhookSchema>
 export type IncomingPaymentCreated = z.infer<typeof incomingPaymentSchema>
@@ -29,7 +30,10 @@ export type GetPaymentDetailsByUrl = z.infer<typeof paymentDetailsSchema>
 export const fakeLoginData = () => {
   return {
     email: faker.internet.email(),
-    password: faker.internet.password()
+    password: faker.internet.password({
+      length: 20,
+      pattern: /[0-9a-zA-Z`!@#$%^&*()_+\-=\[\]{};':"\\|,.<>\/?~]/
+    })
   }
 }
 
@@ -42,7 +46,7 @@ export const mockLogInRequest = (
   }
 })
 
-type SignUpRequest = z.infer<typeof signUpSchema>
+type SignUpRequest = z.infer<typeof signUpBodySchema>
 export const mockSignUpRequest = (
   overrides?: Partial<SignUpRequest['body']>
 ): SignUpRequest => {
